+++ src/main/java/com/jm/online_store/controller/rest/admin/AdminRestController.java
@@ -2,7 +2,7 @@ package com.jm.online_store.controller.rest.admin;
 
 import com.jm.online_store.enums.ResponseOperation;
 import com.jm.online_store.exception.constants.ExceptionConstants;
-import com.jm.online_store.exception.ExceptionEnums;
+import com.jm.online_store.enums.ExceptionEnums;
 import com.jm.online_store.exception.UserServiceException;
 import com.jm.online_store.model.CommonSettings;
 import com.jm.online_store.model.FavouritesGroup;
@@ -99,7 +99,7 @@ public class AdminRestController {
     public ResponseEntity<ResponseDto<UserDto>> getUserInfo(@PathVariable Long id) {
         if (userService.findById(id).isEmpty()) {
             log.debug("User with id: {} not found", id);
-            throw new UserServiceException(ExceptionEnums.USER.getDescription() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
+            throw new UserServiceException(ExceptionEnums.USER.getText() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
         }
         User user = userService.findById(id).get();
         log.debug("User with id: {} found, email is: {}", id, user.getEmail());
@@ -124,7 +124,7 @@ public class AdminRestController {
             userService.deleteByID(id);
         } catch (IllegalArgumentException | EmptyResultDataAccessException | NullPointerException e) {
             log.debug("There is no user with id: {}", id);
-            throw new UserServiceException(ExceptionEnums.USER.getDescription() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
+            throw new UserServiceException(ExceptionEnums.USER.getText() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
         }
         log.debug("User with id: {}, was deleted successfully", id);
         return new ResponseEntity<>(new ResponseDto<>(true, modelMapper.map(userToDelete, UserDto.class)), HttpStatus.OK);
@@ -144,20 +144,20 @@ public class AdminRestController {
     public ResponseEntity<ResponseDto<UserDto>> editUser(@RequestBody User user) {
         if (userService.findById(user.getId()).isEmpty()) {
             log.debug("There are no user with id: {}", user.getId());
-            throw new UserServiceException(ExceptionEnums.USER.getDescription() + ExceptionConstants.NOT_FOUND);
+            throw new UserServiceException(ExceptionEnums.USER.getText() + ExceptionConstants.NOT_FOUND);
         }
         if (ValidationUtils.isNotValidEmail(user.getEmail())) {
             log.debug("Wrong email");
-            throw new UserServiceException(ExceptionEnums.EMAIL.getDescription() + ExceptionConstants.NOT_VALID);
+            throw new UserServiceException(ExceptionEnums.EMAIL.getText() + ExceptionConstants.NOT_VALID);
         }
         if (user.getRoles().size() == 0) {
             log.debug("Roles not selected");
-            throw new UserServiceException(ExceptionEnums.ROLES.getDescription() + ExceptionConstants.NOT_FOUND);
+            throw new UserServiceException(ExceptionEnums.ROLES.getText() + ExceptionConstants.NOT_FOUND);
         }
         if (!userService.findById(user.getId()).get().getEmail().equals(user.getEmail())
                 && userService.isExist(user.getEmail())) {
             log.debug("User with same email already exists");
-            throw new UserServiceException(ExceptionEnums.EMAIL.getDescription() + ExceptionConstants.ALREADY_EXISTS);
+            throw new UserServiceException(ExceptionEnums.EMAIL.getText() + ExceptionConstants.ALREADY_EXISTS);
         }
         userService.updateUserFromAdminPage(user);
         log.debug("Changes to user with id: {} was successfully added", user.getId());
@@ -178,19 +178,19 @@ public class AdminRestController {
     public ResponseEntity<ResponseDto<UserDto>> addNewUser(@RequestBody User newUser) {
         if (ValidationUtils.isNotValidEmail(newUser.getEmail())) {
             log.debug("Wrong email");
-            throw new UserServiceException(ExceptionEnums.EMAIL.getDescription() + ExceptionConstants.NOT_VALID);
+            throw new UserServiceException(ExceptionEnums.EMAIL.getText() + ExceptionConstants.NOT_VALID);
         }
         if (userService.isExist(newUser.getEmail())) {
             log.debug("User with same email already exists");
-            throw new UserServiceException(ExceptionEnums.EMAIL.getDescription() + ExceptionConstants.ALREADY_EXISTS);
+            throw new UserServiceException(ExceptionEnums.EMAIL.getText() + ExceptionConstants.ALREADY_EXISTS);
         }
         if (newUser.getPassword().equals("")) {
             log.debug("Password is empty");
-            throw new UserServiceException(ExceptionEnums.PASSWORD.getDescription() + ExceptionConstants.IS_EMPTY);
+            throw new UserServiceException(ExceptionEnums.PASSWORD.getText() + ExceptionConstants.IS_EMPTY);
         }
         if (newUser.getRoles().size() == 0) {
             log.debug("Roles not selected");
-            throw new UserServiceException(ExceptionEnums.ROLES.getDescription() + ExceptionConstants.IS_EMPTY);
+            throw new UserServiceException(ExceptionEnums.ROLES.getText() + ExceptionConstants.IS_EMPTY);
         }
         userService.addNewUserFromAdmin(newUser);
         User customer = userService.findByEmail(newUser.getEmail()).get();
@@ -220,7 +220,7 @@ public class AdminRestController {
         }
         if (allUsersWithRoleDto.size() == 0) {
             log.debug("There are no users with chosen role in db");
-            throw new UserServiceException(ExceptionEnums.USERS.getDescription() + ExceptionConstants.NOT_FOUND);
+            throw new UserServiceException(ExceptionEnums.USERS.getText() + ExceptionConstants.NOT_FOUND);
         }
         return new ResponseEntity<>(new ResponseDto<>(true, allUsersWithRoleDto), HttpStatus.OK);
     }
