+++ src/main/java/com/jm/online_store/controller/rest/customer/CustomerRestController.java
@@ -3,7 +3,7 @@ package com.jm.online_store.controller.rest.customer;
 import com.jm.online_store.enums.ResponseOperation;
 import com.jm.online_store.exception.CustomerNotFoundException;
 import com.jm.online_store.exception.ExceptionConstants;
-import com.jm.online_store.exception.ExceptionEnums;
+import com.jm.online_store.enums.ExceptionEnums;
 import com.jm.online_store.exception.UserNotFoundException;
 import com.jm.online_store.model.Customer;
 import com.jm.online_store.model.RecentlyViewedProducts;
@@ -112,7 +112,7 @@ public class CustomerRestController {
         }
         catch (IllegalArgumentException | UserNotFoundException e) {
             log.debug("There is no user with id: {}", id);
-            throw new CustomerNotFoundException(ExceptionEnums.CUSTOMER.getDescription() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
+            throw new CustomerNotFoundException(ExceptionEnums.CUSTOMER.getText() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
         }
         User user = userService.findById(id).get();
         log.debug("User with id: {}, was blocked", id);
@@ -136,7 +136,7 @@ public class CustomerRestController {
             customerService.deleteByID(id);
         }
         catch (EmptyResultDataAccessException |IllegalArgumentException | UserNotFoundException e) {
-            throw new CustomerNotFoundException(ExceptionEnums.CUSTOMER.getDescription() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
+            throw new CustomerNotFoundException(ExceptionEnums.CUSTOMER.getText() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
         }
         return new ResponseEntity<>(new ResponseDto<>(true, modelMapper.map(userToDelete, UserDto.class)), HttpStatus.OK);
     }
@@ -155,7 +155,7 @@ public class CustomerRestController {
     public ResponseEntity<ResponseDto<UserDto>> getUserById(@PathVariable("id") Long id) {
         if (userService.findById(id).isEmpty()) {
             log.debug("User with id: {} not found", id);
-            throw new CustomerNotFoundException(ExceptionEnums.CUSTOMER.getDescription() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
+            throw new CustomerNotFoundException(ExceptionEnums.CUSTOMER.getText() + String.format(ExceptionConstants.WITH_SUCH_ID_NOT_FOUND, id));
         }
         log.debug("User with id: {} found", id);
         User user = userService.findUserById(id);
